rename SocialFooter to SiteFooter, more accurate description
