+++ civictechprojects/static/css/styles.scss
@@ -74,7 +74,7 @@
 @import "partials/DonateController";
 @import "partials/ThankYouController";
 @import "partials/Press";
-@import "partials/SocialFooter";
+@import "partials/SiteFooter";
 @import "partials/Newsletter";
 @import "partials/Fonts";
 @import "partials/LoadingMessage";
